requirements.txt
smartapi-python
pandas
numpy
websocket-client
requests
python-dotenv
pytz
pyotp
